requirements.txt
fastapi==0.111.0
langchain==0.2.5
langchain_chroma==0.1.1
langchain_community==0.2.5
langchain_core==0.2.8
langchain_groq==0.1.5
langchain_openai==0.1.8
langchain_text_splitters==0.2.1
python-dotenv==1.0.1
python-telegram-bot==21.4
pymupdf==1.24.9
discord.py==2.4.0